requirements.txt
# Core dependencies
fastapi==0.115.5
uvicorn==0.32.0
# Apenas CPU
torch==2.5.1 --index-url https://download.pytorch.org/whl/cpu
scikit-learn==1.5.2
mlflow==2.17.2
pandas==2.2.3
numpy==2.1.3
yfinance==0.2.49
google-cloud-storage==2.18.2
python-multipart==0.0.18
pydantic==2.9.2
joblib==1.4.2
matplotlib==3.9.2

# Install package in editable mode
-e .